* ralloc.c (relinquish): Use a long for excess space counter to 	handle 64-bit case correctly.

+++ src/ralloc.c
@@ -330,7 +330,7 @@ static void
 relinquish ()
 {
   register heap_ptr h;
-  int excess = 0;
+  long excess = 0;
 
   /* Add the amount of space beyond break_value
      in all heaps which have extend beyond break_value at all.  */
